(+)category id (crud article)

+++ handler/article.go
@@ -69,9 +69,15 @@ func (handler *articleHandler) ArticleStore(c *gin.Context) {
 
 	Title := c.PostForm("Title")
 	Content := c.PostForm("Content")
-	Media, err := c.FormFile("Media")
+	CategoryID := c.PostForm("CategoryID")
+	categoryID, err := strconv.Atoi(CategoryID)
+	if err != nil {
+		categoryID = 0
+		fmt.Println(err)
+		return
+	}
 	Filename := ""
-
+	Media, err := c.FormFile("Media")
 	if err == nil {
 		rand.Seed(time.Now().UnixNano())
 		randNum := rand.Intn(100000)
@@ -86,8 +92,7 @@ func (handler *articleHandler) ArticleStore(c *gin.Context) {
 		}
 
 	}
-	articleRequest := article.ArticleRequest{Title: Title, Media: Filename, Content: Content}
-
+	articleRequest := article.ArticleRequest{Title: Title, Media: Filename, Content: Content, CategoryID: int(categoryID)}
 	article, err := handler.articleService.Create(articleRequest)
 
 	if err != nil {
@@ -116,55 +121,62 @@ func (handler *articleHandler) ArticleUpdate(c *gin.Context) {
 			"errors": err,
 		})
 		return
-	} else if cst.ID == 0 {
+	}
+	if cst.ID == 0 {
 		c.JSON(http.StatusBadRequest, gin.H{
 			"status":  false,
 			"message": "data tidak ditemukan",
 		})
-	} else {
-		Title := c.PostForm("Title")
-		Content := c.PostForm("Content")
-		Media, errmedia := c.FormFile("Media")
-		Filename := ""
+		return
+	}
+	Title := c.PostForm("Title")
+	Content := c.PostForm("Content")
+	CategoryID := c.PostForm("CategoryID")
+	categoryID, err := strconv.Atoi(CategoryID)
+	if err != nil {
+		categoryID = 0
+	}
+	Filename := ""
 
-		if errmedia == nil {
-			rand.Seed(time.Now().UnixNano())
-			randNum := rand.Intn(100000)
-			fileName := strconv.Itoa(randNum) + filepath.Ext(Media.Filename)
+	Media, errmedia := c.FormFile("Media")
+	if errmedia == nil {
+		rand.Seed(time.Now().UnixNano())
+		randNum := rand.Intn(100000)
+		fileName := strconv.Itoa(randNum) + filepath.Ext(Media.Filename)
 
-			Filename = fmt.Sprintf("uploads/codelite_%s", fileName)
-			if err := c.SaveUploadedFile(Media, Filename); err != nil {
-				c.JSON(500, gin.H{
-					"error": "Failed to save media",
-				})
-				return
-			} else {
+		Filename = fmt.Sprintf("uploads/codelite_%s", fileName)
+		if err := c.SaveUploadedFile(Media, Filename); err != nil {
+			c.JSON(500, gin.H{
+				"error": "Failed to save media",
+			})
+			return
+		} else {
+			if cst.Media != "" {
 				err := os.Remove(cst.Media)
 				if err != nil {
 					fmt.Println("Error deleting file:", err)
-					return
 				}
 			}
-
 		}
 
-		articleRequest := article.ArticleUpdateRequest{Title: Title, Media: Filename, Content: Content}
-
-		article, err := handler.articleService.Update(id, articleRequest)
-
-		if err != nil {
-			c.JSON(http.StatusBadRequest, gin.H{
-				"errors": err,
-			})
-			return
-		}
-
-		c.JSON(http.StatusOK, gin.H{
-			"status":  true,
-			"message": "Data tersimpan",
-			"data":    article,
-		})
 	}
+
+	articleRequest := article.ArticleUpdateRequest{Title: Title, Media: Filename, Content: Content, CategoryID: int(categoryID)}
+
+	article, err := handler.articleService.Update(id, articleRequest)
+
+	if err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"errors": err,
+		})
+		return
+	}
+
+	c.JSON(http.StatusOK, gin.H{
+		"status":  true,
+		"message": "Data tersimpan",
+		"data":    article,
+	})
 
 }
 
@@ -212,11 +224,12 @@ func (handler *articleHandler) ArticleDelete(c *gin.Context) {
 
 func responseArticle(cst article.Article) article.ArticleResponse {
 	return article.ArticleResponse{
-		ID:        cst.ID,
-		Title:     cst.Title,
-		Content:   cst.Content,
-		Media:     cst.Media,
-		CreatedAt: cst.CreatedAt,
-		UpdatedAt: cst.UpdatedAt,
+		ID:         cst.ID,
+		Title:      cst.Title,
+		Content:    cst.Content,
+		Media:      cst.Media,
+		CategoryID: cst.CategoryID,
+		CreatedAt:  cst.CreatedAt,
+		UpdatedAt:  cst.UpdatedAt,
 	}
 }
